Changed tests to be easier to run
