+++ src/shimoku_api_python/resources/report.py
@@ -226,6 +226,9 @@ class Report(Resource):
             elif db_resource['reportType'] == 'FORM':
                 from .reports.charts.input_form import InputForm
                 return InputForm(parent=parent, uuid=uuid, db_resource=db_resource)
+            elif db_resource['reportType'] == 'FILTERDATASET':
+                from .reports.filter_data_set import FilterDataSet
+                return FilterDataSet(parent=parent, uuid=uuid, db_resource=db_resource)
             else:
                 raise ValueError(f'Unknown report type {db_resource["reportType"]}')
         else:
